Move compatibility import for `tomllib` to tests/
This is not used outside of the test suite.
