+++ tests/compat.py
