+++ tests/test_call_hooks.py
@@ -16,7 +16,7 @@ from pyproject_hooks import (
     UnsupportedOperation,
     default_subprocess_runner,
 )
-from pyproject_hooks._compat import tomllib
+from tests.compat import tomllib
 
 SAMPLES_DIR = pjoin(dirname(abspath(__file__)), 'samples')
 BUILDSYS_PKGS = pjoin(SAMPLES_DIR, 'buildsys_pkgs')
