+++ tests/test_inplace_hooks.py
@@ -5,7 +5,7 @@ import pytest
 from testpath import modified_env
 
 from pyproject_hooks import BackendInvalid, BuildBackendHookCaller
-from pyproject_hooks._compat import tomllib
+from tests.compat import tomllib
 
 SAMPLES_DIR = pjoin(dirname(abspath(__file__)), 'samples')
 BUILDSYS_PKGS = pjoin(SAMPLES_DIR, 'buildsys_pkgs')
